Refactor booking

+++ src/BS.Application/Services/BookingService.cs
@@ -33,7 +33,7 @@ public class BookingService : IBookingService
         _spotRepository = spotRepository;
     }
 
-    public async Task Book(BookingDto dto)
+    public async Task BookAsync(BookingDto dto)
     {
         if (dto is null)
             throw new ArgumentNullException(nameof(dto));
@@ -41,9 +41,6 @@ public class BookingService : IBookingService
         var user = await _userManager.FindByIdAsync(dto.UserId);
         var spot = await _spotRepository
             .GetByIdWithResourceAsync(dto.SpotId);
-
-        if (spot == null)
-            throw new InvalidOperationException("Spot not found");
 
         if (user == null || spot == null)
         {
@@ -58,7 +55,6 @@ public class BookingService : IBookingService
         var booking = Booking.Create(spot, user.Id, dto.StartTime, dto.EndTime);
         
         spot.IsActive = false;
-        await _spotGenericRepository.SaveChangesAsync();
 
         await _repository.Create(booking);
         await _repository.SaveChangesAsync();
@@ -67,17 +63,27 @@ public class BookingService : IBookingService
     public async Task<BookingDto?> FindBookingById(int id)
     {
         var booking = await _repository.GetById(id);
+        if (booking is null)
+            return null;
+        
         return _mapper.Map<BookingDto>(booking);
     }
 
     public async Task<List<BookingDto>?> FindBookingByUserId(string id)
     {
         var booking = await _bookingRepository.GetByUserId(id);
+        if (booking is null)
+            return null;
+        
         return _mapper.Map<List<BookingDto>>(booking);
     }
 
     public async Task DeleteBooking(int id)
     {
+        var booking = await _repository.GetById(id);
+        if (booking is null)
+            throw new NullReferenceException("Booking not found");
+        
         await _repository.Delete(id);
         await _repository.SaveChangesAsync();
     }
